implementing sanctum middleware over de request routes

+++ routes/api.php
@@ -23,10 +23,10 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::get('/weather', [WeatherRequestController::class, 'index']); // apply route groups and validate numeric id
-Route::post('/weather', [WeatherRequestController::class, 'store']);
-Route::get('/weather/{id}', [WeatherRequestController::class, 'show']); // apply route groups and validate numeric id
-Route::delete('/weather/{id}', [WeatherRequestController::class, 'destroy']); // apply route groups and validate numeric id
+Route::middleware('auth:sanctum')->get('/weather', [WeatherRequestController::class, 'index']); // apply route groups and validate numeric id
+Route::middleware('auth:sanctum')->post('/weather', [WeatherRequestController::class, 'store']);
+Route::middleware('auth:sanctum')->get('/weather/{id}', [WeatherRequestController::class, 'show']); // apply route groups and validate numeric id
+Route::middleware('auth:sanctum')->delete('/weather/{id}', [WeatherRequestController::class, 'destroy']); // apply route groups and validate numeric id
 
 Route::post('/register', [UserController::class, 'store']);
